requirements.txt
aniso8601==8.0.0
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
dask==2.9.0
decorator==4.4.1
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==20.0.4
guppy3==3.0.9
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.19.1
Pillow==6.2.1
psutil==5.6.7
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.14.2
scipy==1.3.3
six==1.13.0
toolz==0.10.0
Werkzeug==0.16.0
